+++ src/scripts/core/brt-table-results.js
@@ -18,7 +18,7 @@ export class BetterResults {
     // START PATCH 2024-03-02
     // this.currencyData = await this._generateCurrency(currencyString);
     const currencyString = table.getFlag(CONSTANTS.MODULE_ID, CONSTANTS.FLAGS.LOOT_CURRENCY_STRING_KEY);
-    this.currencyData = await ItemPilesHelpers.getCurrenciesSimpleFromString(currencyString);
+    this.currencyData = await ItemPilesHelpers.retrieveCurrenciesSimpleFromString(currencyString);
     // END PATCH 2024-03-02
     // START PATCH 2024-03-02
     /*