+++ src/scripts/tables/harvest/harvest-chat-card.js
@@ -48,6 +48,14 @@ export class HarvestChatCard {
 
       let isResultHidden = false;
       if (getProperty(result, `flags.${CONSTANTS.MODULE_ID}.${CONSTANTS.FLAGS.GENERIC_RESULT_HIDDEN_TABLE}`)) {
+        if (
+          !getProperty(
+            result,
+            `flags.${CONSTANTS.MODULE_ID}.${CONSTANTS.FLAGS.GENERIC_RESULT_SHOW_HIDDEN_RESULT_ON_CHAT}`
+          )
+        ) {
+          continue;
+        }
         // customResultNameHidden = CONSTANTS.DEFAULT_HIDDEN_RESULT_TEXT;
         // customResultImgHidden = CONSTANTS.DEFAULT_HIDDEN_RESULT_IMAGE;
         isResultHidden = true;