+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 TODO 
 - Add confirm dialog on reset button ???
+- Add flags for cost/price like already done for the quantity is possible now with item piles support
+
 
 ### 2.0.34